Add files via upload
In the project that I am showing the code that is going to allow Chada Tech clients to be able to view a 12- and a 24-hour clock on their website rather than just the standard 12-hour clock. So, the part of the code that I found to be most difficult when writing the code to make what the company wanted was the actual writing the code itself. I had a lot of help from tutors. They were a big help to overcome the obstacles that I had faced while writing the code. I think that every coding language is special in its own way and that once you learn it then you can transfer it to the real job. The language is easier to read when you know the language and what it would print out in the end.